+++ modules/mod_proc/mod_proc.c
@@ -98,7 +98,7 @@ static void _modproc_kill_all()
 
 /* ----------------------------------------------------------------- */
 
-static int modproc_exit_0( INSTANCE * my, int * params )
+static int modproc_exit_0( INSTANCE * my, intptr_t * params )
 {
     exit_value = 0;
     must_exit = 1 ;
@@ -108,7 +108,7 @@ static int modproc_exit_0( INSTANCE * my, int * params )
 
 /* ----------------------------------------------------------------- */
 
-static int modproc_exit_1( INSTANCE * my, int * params )
+static int modproc_exit_1( INSTANCE * my, intptr_t * params )
 {
     printf( "%s\n", string_get( params[0] ) );
     fflush( stdout );
@@ -122,7 +122,7 @@ static int modproc_exit_1( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int modproc_exit( INSTANCE * my, int * params )
+static int modproc_exit( INSTANCE * my, intptr_t * params )
 {
     _modproc_kill_all();
 
@@ -138,7 +138,7 @@ static int modproc_exit( INSTANCE * my, int * params )
 
 /* ----------------------------------------------------------------- */
 
-static int modproc_running( INSTANCE * my, int * params )
+static int modproc_running( INSTANCE * my, intptr_t * params )
 {
     INSTANCE * i, * ctx;
 
@@ -162,7 +162,7 @@ static int modproc_running( INSTANCE * my, int * params )
 
 /* ----------------------------------------------------------------- */
 
-static int modproc_signal( INSTANCE * my, int * params )
+static int modproc_signal( INSTANCE * my, intptr_t * params )
 {
     INSTANCE * i, * ctx;
     int fake_params[2] ;
@@ -275,7 +275,7 @@ static int modproc_signal( INSTANCE * my, int * params )
 
 /* ----------------------------------------------------------------- */
 
-static int modproc_signal_action( INSTANCE * my, int * params )
+static int modproc_signal_action( INSTANCE * my, intptr_t * params )
 {
     int action = params[1];
 
@@ -393,7 +393,7 @@ static int modproc_signal_action( INSTANCE * my, int * params )
 
 /* ----------------------------------------------------------------- */
 
-static int modproc_signal_action3( INSTANCE * my, int * params )
+static int modproc_signal_action3( INSTANCE * my, intptr_t * params )
 {
     INSTANCE * i, * ctx ;
 
@@ -425,7 +425,7 @@ static int modproc_signal_action3( INSTANCE * my, int * params )
 
 /* ----------------------------------------------------------------- */
 
-static int modproc_let_me_alone( INSTANCE * my, int * params )
+static int modproc_let_me_alone( INSTANCE * my, intptr_t * params )
 {
     INSTANCE * i = first_instance ;
 
@@ -441,7 +441,7 @@ static int modproc_let_me_alone( INSTANCE * my, int * params )
 
 /* ----------------------------------------------------------------- */
 
-static int modproc_get_id( INSTANCE * my, int * params )
+static int modproc_get_id( INSTANCE * my, intptr_t * params )
 {
     INSTANCE * ptr = first_instance, ** ctx ;
 
@@ -496,7 +496,7 @@ static int modproc_get_id( INSTANCE * my, int * params )
 
 /* ----------------------------------------------------------------- */
 
-static int modproc_get_status( INSTANCE * my, int * params )
+static int modproc_get_status( INSTANCE * my, intptr_t * params )
 {
     INSTANCE * i ;
     if ( !params[0] || !( i = instance_get( params[0] ) ) ) return 0;
